Fix filter settings
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-2.0.9@223 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/audio/mmsystem/mmsystem.cpp
@@ -121,7 +121,7 @@ short *Audio_MMSystem::open (AudioConfig &cfg, const char *)
             }
 
             /* Lock mixing block memory: */
-            if ( (blocks[i] = (float *)GlobalLock(blockHandles[i])) == NULL )
+            if ( (blocks[i] = (short *)GlobalLock(blockHandles[i])) == NULL )
             {
                 _errorString = "MMSYSTEM ERROR: Can't lock global memory.";
                 goto Audio_MMSystem_openError;
